Revert "[feat] Changed entities and introduced TransferService"
