--- src/main/java/edu/aseca/bags/domain/WalletRepository.java
@@ -1,13 +1,0 @@
-package edu.aseca.bags.domain;
-
-import edu.aseca.bags.domain.email.Email;
-import edu.aseca.bags.domain.wallet.Wallet;
-import java.util.Optional;
-import java.util.UUID;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface WalletRepository extends JpaRepository<Wallet, UUID> {
-	Optional<Wallet> findByEmail(Email fromEmail);
-}
