--- src/main/java/edu/aseca/bags/domain/transaction/InsufficientFundsException.java
@@ -1,5 +1,0 @@
-package edu.aseca.bags.domain.transaction;
-
-public class InsufficientFundsException extends Exception {
-
-}
