--- src/main/java/edu/aseca/bags/domain/transaction/TransactionType.java
@@ -1,5 +1,0 @@
-package edu.aseca.bags.domain.transaction;
-
-public enum TransactionType {
-	TRANSFER, EXTRACTION, DEPOSIT
-}
