--- src/main/java/edu/aseca/bags/domain/wallet/WalletNotFoundException.java
@@ -1,4 +1,0 @@
-package edu.aseca.bags.domain.wallet;
-
-public class WalletNotFoundException extends Exception {
-}
