+++ src/main/java/edu/aseca/bags/email/Email.java
@@ -1,9 +1,7 @@
-package edu.aseca.bags.domain.email;
+package edu.aseca.bags.email;
 
-import jakarta.persistence.Embeddable;
 import java.util.regex.Pattern;
 
-@Embeddable
 public record Email(String address) {
 	private static final Pattern EMAIL_PATTERN = Pattern.compile("^(?!.*\\.\\.)[A-Z0-9._%+-]+@[A-Z0-9.-]+\\.[A-Z]{2,}$",
 			Pattern.CASE_INSENSITIVE);
