+++ src/main/resources/application.properties
@@ -3,6 +3,3 @@ server.port=${APP_PORT}
 spring.datasource.url=jdbc:postgresql://${DB_IP}:${DB_PORT}/${DB_NAME}
 spring.datasource.username=${POSTGRES_USER}
 spring.datasource.password=${POSTGRES_PASSWORD}
-
-spring.jpa.hibernate.ddl-auto=update
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
