+++ src/test/resources/application-test.properties
@@ -15,6 +15,3 @@ spring.jpa.hibernate.ddl-auto=create-drop
 server.port=8080
 
 logging.level.org.springframework.boot.context.properties=DEBUG
-
-logging.level.org.hibernate.SQL=DEBUG
-logging.level.org.hibernate.tool.hbm2ddl=DEBUG
